requirements.txt

# Core dependencies
einops==0.8.1
numpy==2.2.0
pillow==12.0.0
diffusers==0.35.2
safetensors==0.7.0
tokenizers==0.22.0
transformers[accelerate,tiktoken]==4.57.1
huggingface_hub[cli]
loguru>=0.7.3

# PyTorch with CUDA 12.8 support (install separately)
# torch==2.8.0
# torchvision==0.23.0  
# torchaudio==2.8.0
# Install with: pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cu128

# Performance optimizations (recommended for up to 3x faster inference)
# flashinfer-python==0.5.0
# Install with: pip install flashinfer-python==0.5.0

# Interactive demo
gradio>=4.21.0



